Move navbar components to separate files
