+++ src/ui/components/Layout.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Wrapper, WrapperVariant } from './Wrapper';
-import NavBar from '../widgets/NavBar';
+import NavBar from '@minos/ui/widgets/navbar/NavBar';
 import { Center } from '@chakra-ui/react';
 
 interface LayoutProps {
